requirements.txt
pyrogram==2.0.106
python-dotenv==1.0.0
rich==13.7.0
flet